상품_엘라스틱서치_로그스테시로_데이터_불러온거_검색_확인_상품사이즈_한번에_저장_태그_수정중 feature#2-13

+++ src/main/java/com/kj/products/product/repository/ProductRepositoryCustom.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public interface ProductRepositoryCustom {
     int findByMaxProductId();
 
-    Optional<Product> findByProductId(int no);
+    Optional<Product> findByProductId(long no);
 
     PageImpl<ProductListDto> findListProducPage(Pageable pageable, ProductSearchCondotion productSearchCondotion);
